--- python/ray/train/backends/__init__.py
@@ -1,13 +1,0 @@
-from ray.train.backends.backend import BackendConfig
-from ray.train.backends.horovod import HorovodConfig
-from ray.train.backends.tensorflow import TensorflowConfig
-
-# Backwards Compatiblity
-from ray.train.torch import TorchConfig
-
-__all__ = [
-    "BackendConfig",
-    "HorovodConfig",
-    "TensorflowConfig",
-    "TorchConfig",
-]
